remove header for plugin SiPixelRawToDigi
